pdfcb/requirements.txt
flask
pymongo
openai
pdfminer.six
python-dotenv
numpy